tests and docker file for scrum poker

+++ dev projects/ScrumPoker/ScrumPoker/Service/IDevQAService.cs
@@ -4,6 +4,6 @@ public interface IDevQAService
 {
     Task<Game> JoinGameAsync(string gameCode, int userId);
     Task<Vote> CastVoteAsync(string gameCode, Vote vote, int userId);
-    Task<int> GetDevVoteAsync(string gameCode);
-    Task<int> GetQAVoteAsync(string gameCode);
+    Task<double> GetDevVoteAsync(string gameCode);
+    Task<double> GetQAVoteAsync(string gameCode);
 }
